requirements.txt
schedule
dotenv
load_dotenv
nats-py
jsonschema==4.19.0
